requirements.txt
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.3
fsspec==2024.10.0
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.17.0
SQLAlchemy==2.0.36
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.49
ultralytics-thop==2.0.13
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
